api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
transformers==4.38.2
safetensors==0.4.2
torch==2.2.1
pydantic==2.6.3
langchain==0.1.9
langchain-community==0.0.24
faiss-cpu==1.7.4
sentence-transformers==2.5.1
python-multipart==0.0.9
